MDEV-27524: Incorrect binlogs after Galera SST using rsync and mariabackup
This commit adds correct handling of binlogs for SST using rsync
or mariabackup. Before this fix, binlogs were handled incorrectly -
- only one (last) binary log file was transferred during SST, which
then led to various failures, for example, when trying to list all
events from the binary log. These bugs were long masked by flaws in
the very primitive binlogs handling code in the SST scripts, which
caused binary logs files to be erased or not added to the index on
the joiner node. Now the correct transfer of all binlogs (not just
the last of the binary log file) has been added both for the rsync
(at the script level) and for the mariabackup (at the level of the
main utility code).

Also this commit adds a new sst_max_binlogs=<n> parameter, which
can be located in the [sst] section or in the [xtrabackup] section
(historically and for mariabackup only), or in one of the server
sections. This parameter specifies the number of binary logs to
be sent to the joiner node during SST. This option is added for
compatibility with old SST scripting behavior, which can be emulated
by setting the sst_max_binlogs=1 (although in general this can cause
problems for the reasons described above). In addition, setting
the sst_max_binlogs=0 mode can be used to suppress the transmission
of binary logs to the joiner nodes (although sometimes a single
current binary log can still be transmitted to the joiner even with
sst_max_binlogs=0, because this sometimes necessary in modes that
involve the use of GTIDs with Galera).

In addition, this commit contains the following enhancements:

 1) Added tests for mtr, which check the correct work with binlogs
    after SST (using rsync and mariabackup);
 2) Added correct handling of slashes at the end of all paths
    that the SST script receives as parameters;
 3) Improved parsing code for mysqld_args parameters, now it
    correctly processes the sequence "--" after the name of
    the one-letter option;
 4) Checking the secret code during joiner authentication is
    made independent of bash (as a unix shell);
 5) All directories that are necessary for the correct placement
    of various logs are created in advance (before running
    mariabackup on joiner);
 6) Removal of old binlogs on joiner is done by binlogs index
    (if it exists) and not by pattern with current binlog name;
 7) Paths for placing binlogs are correctly processed if they
    are set relative to datadir;
 8) Scripts are made even more resistant to spaces in filenames
    (now for binlogs);
 9) In case of errors, scripts should now always end with
    an exit code other than zero;
10) SST scripts also now correctly create a tar file with
    binlogs if the paths to them in the index file are specified
    as absolute paths, even if they do not match the last one.